Added DevExpress. Create models for calculation variation series.

+++ ReliabilityAssessmentApp/ReliabilityAssessmentLibrary/StandardPointEstimate.cs
@@ -15,6 +15,9 @@ namespace ReliabilityAssessmentLibrary
             this._dataModel = dataModel ?? throw new ArgumentException(nameof(dataModel));
         }
 
+        /// <summary>
+        /// Найдем оценку математического ожидания 
+        /// </summary>
         public double GetMathematicalExpectation()
         {
             var mathExpectation = this._dataModel.ValueList.Sum()
@@ -24,16 +27,22 @@ namespace ReliabilityAssessmentLibrary
             return mathExpectation;
         }
 
+        /// <summary>
+        /// Расcчитаем оценку дисперсии
+        /// </summary>
         public double GetEstimationVariance()
         {
-            var estimateVariance = 1 / (this._dataModel.ValueList.Count() - 1)
-                * this._dataModel.ValueList.Select(x => Math.Pow(x, 2)).Sum()
-                - this._dataModel.ValueList.Count * Math.Pow(this._dataModel.MathExpectation, 2);
+            var estimateVariance = 1 / ((double)this._dataModel.ValueList.Count() - 1)
+                * (this._dataModel.ValueList.Select(x => Math.Pow(x, 2)).Sum()
+                - this._dataModel.ValueList.Count * Math.Pow(this._dataModel.MathExpectation, 2));
 
             this._dataModel.SetDispersion(estimateVariance);
             return estimateVariance;
         }
 
+        /// <summary>
+        /// Рассчитаем среднее квадратичное отклонение
+        /// </summary>
         public double GetMeanSquareDeviation()
         {
             var meanSquareDeviation = Math.Sqrt(this._dataModel.Dispersion);
@@ -41,6 +50,9 @@ namespace ReliabilityAssessmentLibrary
             return meanSquareDeviation;
         }
 
+        /// <summary>
+        /// Рассчитаем коэффициент вариации
+        /// </summary>
         public double CoefficientVariation()
         {
             var coeffVariation = this._dataModel.SquareDeviation / this._dataModel.MathExpectation;
